Flytt teste fra feilsøking til eksisterende test
